% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate BMI}
\usage{
calculate(weight_kg, height_m)
}
\arguments{
\item{weight_kg}{Numeric. Weight in kilograms.}

\item{height_m}{Numeric. Height in meters.}
}
\value{
The calculated BMI (numeric).
}
\description{
This function calculates the BMI given weight in kg and height in meters.
}
\examples{
calculate(70, 1.80)
}
