% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertInch}
\alias{convertInch}
\title{Convert inches to meters}
\usage{
convertInch(height_in)
}
\arguments{
\item{height_in}{Numeric. Height in inches.}
}
\value{
The height in meters.
}
\description{
This function converts height from inches to meters.
}
\examples{
convertInch(70)
}
