% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertLbs}
\alias{convertLbs}
\title{Convert lbs to kg}
\usage{
convertLbs(weight_lbs)
}
\arguments{
\item{weight_lbs}{Numeric. Weight in lbs.}
}
\value{
The weight in kilograms.
}
\description{
This function converts weight from lbs to kg.
}
\examples{
convertLbs(154.32)
}
