% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\title{Interpret BMI}
\usage{
interpret(bmi)
}
\arguments{
\item{bmi}{Numeric. The BMI to interpret.}
}
\value{
The interpretation of the BMI as a string.
}
\description{
This function provides a verbal interpretation of the BMI value.
}
\examples{
interpret(22)
}
